pyproject.toml
[tool.poetry]
name = "chapter4"
version = "0.1.0"
description = ""
authors = ["nishishimakenya <nijijijima0960@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
djangorestframework = "^3.14.0"
flake8 = "^6.0.0"
black = "^23.7.0"
pygments = "^2.15.1"
django-filter = "^23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
